Lektura
It seems to me that something is missing in the first part of the sentence: This is so only one css file is generated, and when a file is not imported, it will not be unnecessarily compiled.
But I can't tell what from the context. Could you help me out and explain the meaning of the sentence please?

Also, please review this sentence, I added "you are", but I'm not sure if I made the right choice:
If you are using LibSass, it might not work (currently it doesn't) so files need to be imported individually.